+++ src/main/java/com/example/financeapp/modules/depot/responses/DepotResponse.java
@@ -1,4 +1,4 @@
-package com.example.financeapp.modules.depot.response;
+package com.example.financeapp.modules.depot.responses;
 
 import lombok.Getter;
 import lombok.Setter;
