+++ src/main/java/com/example/financeapp/modules/orderbook/responses/OrderResponse.java
@@ -1,4 +1,4 @@
-package com.example.financeapp.modules.orderbook.response;
+package com.example.financeapp.modules.orderbook.responses;
 
 import com.example.financeapp.modules.orderbook.OrderEnums;
 import lombok.Getter;
