+++ src/main/java/com/example/financeapp/modules/share/ShareController.java
@@ -1,6 +1,6 @@
 package com.example.financeapp.modules.share;
 
-import com.example.financeapp.modules.share.response.PerformanceResponse;
+import com.example.financeapp.modules.share.responses.PerformanceResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
